+++ tests/Unit/Attribute/ApiController/GetEntityTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Attribute\Api;
+namespace Ifrost\ApiFoundation\Tests\Unit\Attribute\ApiController;
 
-use Ifrost\ApiFoundation\Attribute\Api;
+use Ifrost\ApiFoundation\Attribute\ApiController;
 use Ifrost\ApiFoundation\Tests\Variant\Entity\Product;
 use PHPUnit\Framework\TestCase;
 
@@ -13,7 +13,7 @@ class GetEntityTest extends TestCase
     public function testShouldReturnEntityAsString()
     {
         // Given
-        $attribute = new Api(Product::class, 'products');
+        $attribute = new ApiController(Product::class, 'products');
 
         // When & Then
         $this->assertEquals(Product::class, $attribute->getEntity());