+++ tests/Variant/Controller/InvalidController.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Ifrost\ApiFoundation\Tests\Variant\Controller;
 
-use Ifrost\ApiFoundation\Attribute\Api;
+use Ifrost\ApiFoundation\Attribute\ApiController;
 use Ifrost\ApiFoundation\Tests\Variant\Sample;
 
-#[Api(entity: Sample::class)]
+#[ApiController(entity: Sample::class)]
 class InvalidController extends ApiControllerVariant
 {
 }