+++ tests/Variant/Controller/ProductControllerWithOverwrittenAction.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Ifrost\ApiFoundation\Tests\Variant\Controller;
 
-use Ifrost\ApiFoundation\Attribute\Api;
+use Ifrost\ApiFoundation\Attribute\ApiController;
 use Ifrost\ApiFoundation\Tests\Variant\Entity\Product;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Api(entity: Product::class, path: 'products')]
+#[ApiController(entity: Product::class, path: 'products')]
 class ProductControllerWithOverwrittenAction extends ApiControllerVariant
 {
     #[Route('/create_products', name: 'products_create', methods: ['POST'])]